Move releases methods to new file
